Fix client fallback module imports
