+++ crates/core/src/client/mod.rs
@@ -85,6 +85,10 @@
 //! - [`crate::version`] for the canonical version banner shared with the CLI.
 //! - [`rsync_engine::local_copy`] for the transfer plan executed by this module.
 
+mod fallback;
+
+pub use self::fallback::{RemoteFallbackArgs, RemoteFallbackContext, run_remote_transfer_fallback};
+
 use std::collections::HashMap;
 use std::env::VarError;
 use std::ffi::{OsStr, OsString};
@@ -96,14 +100,15 @@ use std::net::{SocketAddr, TcpStream};
 use std::num::NonZeroU64;
 use std::path::{Path, PathBuf};
 use std::process::{Child, ChildStdin, ChildStdout, Command, Stdio};
-use std::sync::{
-    Arc,
-    mpsc::{self, Sender},
-};
-use std::thread;
+use std::sync::Arc;
 use std::time::{Duration, Instant, SystemTime};
 use std::{env, error::Error};
 
+use crate::{
+    bandwidth::{self, BandwidthLimiter, BandwidthParseError},
+    message::{Message, Role},
+    rsync_error,
+};
 use base64::Engine as _;
 use base64::engine::general_purpose::{STANDARD, STANDARD_NO_PAD};
 use rsync_checksums::strong::Md5;
@@ -129,14 +134,6 @@ use rsync_transport::negotiate_legacy_daemon_session;
 use socket2::{Domain, Protocol, SockAddr, Socket, Type};
 #[cfg(test)]
 use std::cell::RefCell;
-use tempfile::NamedTempFile;
-
-use crate::{
-    bandwidth::{self, BandwidthLimiter, BandwidthParseError},
-    fallback::{CLIENT_FALLBACK_ENV, FallbackOverride, fallback_override},
-    message::{Message, Role},
-    rsync_error,
-};
 
 #[cfg(unix)]
 use std::os::unix::ffi::{OsStrExt, OsStringExt};
@@ -2944,1074 +2941,6 @@ impl FallbackSummary {
     }
 }
 
-/// Arguments used to spawn the legacy `rsync` binary when remote operands are present.
-///
-/// The fallback path preserves the command-line semantics of upstream rsync while the
-/// native protocol engine is completed. Higher level consumers such as the CLI build
-/// this structure from parsed flags before handing control to
-/// [`run_remote_transfer_fallback`].
-#[derive(Clone, Debug)]
-pub struct RemoteFallbackArgs {
-    /// Enables `--dry-run`.
-    pub dry_run: bool,
-    /// Enables `--list-only`.
-    pub list_only: bool,
-    /// Supplies the remote shell command forwarded via `-e`/`--rsh`.
-    pub remote_shell: Option<OsString>,
-    /// Additional options forwarded to the remote rsync invocation via `--remote-option`/`-M`.
-    pub remote_options: Vec<OsString>,
-    /// Optional command executed to reach rsync:// daemons.
-    pub connect_program: Option<OsString>,
-    /// Default daemon port forwarded via `--port` when contacting rsync:// daemons.
-    #[doc(alias = "--port")]
-    pub port: Option<u16>,
-    /// Optional bind address forwarded via `--address`.
-    pub bind_address: Option<OsString>,
-    /// Controls whether remote shell arguments are protected from expansion.
-    ///
-    /// When `Some(true)` the fallback command receives `--protect-args`,
-    /// while `Some(false)` forwards `--no-protect-args`. A `None` value keeps
-    /// rsync's default behaviour.
-    pub protect_args: Option<bool>,
-    /// Optional `--human-readable` level forwarded to the fallback binary.
-    pub human_readable: Option<HumanReadableMode>,
-    /// Enables archive mode (`-a`).
-    pub archive: bool,
-    /// Enables `--delete`.
-    pub delete: bool,
-    /// Selects the deletion timing to forward to the fallback binary.
-    pub delete_mode: DeleteMode,
-    /// Enables `--delete-excluded`.
-    pub delete_excluded: bool,
-    /// Limits deletions via `--max-delete`.
-    pub max_delete: Option<u64>,
-    /// Skips files smaller than the provided size via `--min-size`.
-    pub min_size: Option<OsString>,
-    /// Skips files larger than the provided size via `--max-size`.
-    pub max_size: Option<OsString>,
-    /// Enables `--checksum`.
-    pub checksum: bool,
-    /// Optional strong checksum selection forwarded via `--checksum-choice`.
-    pub checksum_choice: Option<OsString>,
-    /// Optional checksum seed forwarded via `--checksum-seed`.
-    pub checksum_seed: Option<u32>,
-    /// Enables `--size-only`.
-    pub size_only: bool,
-    /// Enables `--ignore-existing`.
-    pub ignore_existing: bool,
-    /// Enables `--ignore-missing-args`.
-    pub ignore_missing_args: bool,
-    /// Enables `--update`.
-    pub update: bool,
-    /// Optional `--modify-window` tolerance forwarded to the fallback binary.
-    pub modify_window: Option<u64>,
-    /// Enables `--compress`.
-    pub compress: bool,
-    /// Enables `--no-compress` when `true` and compression is otherwise disabled.
-    pub compress_disabled: bool,
-    /// Optional compression level forwarded via `--compress-level`.
-    pub compress_level: Option<OsString>,
-    /// Optional suffix list forwarded via `--skip-compress`.
-    pub skip_compress: Option<OsString>,
-    /// Optional ownership override forwarded via `--chown`.
-    pub chown: Option<OsString>,
-    /// Optional `--owner`/`--no-owner` toggle.
-    pub owner: Option<bool>,
-    /// Optional `--group`/`--no-group` toggle.
-    pub group: Option<bool>,
-    /// Repeated `--chmod` specifications forwarded to the fallback binary.
-    pub chmod: Vec<OsString>,
-    /// Optional `--perms`/`--no-perms` toggle.
-    pub perms: Option<bool>,
-    /// Optional `--super`/`--no-super` toggle.
-    pub super_mode: Option<bool>,
-    /// Optional `--times`/`--no-times` toggle.
-    pub times: Option<bool>,
-    /// Optional `--omit-dir-times`/`--no-omit-dir-times` toggle.
-    pub omit_dir_times: Option<bool>,
-    /// Optional `--omit-link-times`/`--no-omit-link-times` toggle.
-    pub omit_link_times: Option<bool>,
-    /// Optional `--numeric-ids`/`--no-numeric-ids` toggle.
-    pub numeric_ids: Option<bool>,
-    /// Optional `--hard-links`/`--no-hard-links` toggle.
-    pub hard_links: Option<bool>,
-    /// Optional `--copy-links`/`--no-copy-links` toggle.
-    pub copy_links: Option<bool>,
-    /// Enables `--copy-dirlinks` when `true`.
-    pub copy_dirlinks: bool,
-    /// Optional `--copy-unsafe-links`/`--no-copy-unsafe-links` toggle.
-    pub copy_unsafe_links: Option<bool>,
-    /// Optional `--keep-dirlinks`/`--no-keep-dirlinks` toggle.
-    pub keep_dirlinks: Option<bool>,
-    /// Enables `--safe-links` when `true`.
-    pub safe_links: bool,
-    /// Optional `--sparse`/`--no-sparse` toggle.
-    pub sparse: Option<bool>,
-    /// Optional `--devices`/`--no-devices` toggle.
-    pub devices: Option<bool>,
-    /// Optional `--specials`/`--no-specials` toggle.
-    pub specials: Option<bool>,
-    /// Optional `--relative`/`--no-relative` toggle.
-    pub relative: Option<bool>,
-    /// Optional `--one-file-system`/`--no-one-file-system` toggle.
-    pub one_file_system: Option<bool>,
-    /// Optional `--implied-dirs`/`--no-implied-dirs` toggle.
-    pub implied_dirs: Option<bool>,
-    /// Enables `--mkpath`.
-    pub mkpath: bool,
-    /// Controls pruning of empty directories via `--prune-empty-dirs`.
-    pub prune_empty_dirs: Option<bool>,
-    /// Verbosity level translated into repeated `-v` flags.
-    pub verbosity: u8,
-    /// Enables `--progress`.
-    pub progress: bool,
-    /// Enables `--stats`.
-    pub stats: bool,
-    /// Enables `--itemize-changes` on the fallback command line.
-    pub itemize_changes: bool,
-    /// Enables `--partial`.
-    pub partial: bool,
-    /// Enables `--preallocate`.
-    pub preallocate: bool,
-    /// Enables `--delay-updates`.
-    pub delay_updates: bool,
-    /// Optional directory forwarded via `--partial-dir`.
-    pub partial_dir: Option<PathBuf>,
-    /// Optional directory forwarded via `--temp-dir`.
-    pub temp_directory: Option<PathBuf>,
-    /// Enables `--backup`.
-    pub backup: bool,
-    /// Optional directory forwarded via `--backup-dir`.
-    pub backup_dir: Option<PathBuf>,
-    /// Optional suffix forwarded via `--suffix`.
-    pub backup_suffix: Option<OsString>,
-    /// Directories forwarded via repeated `--link-dest` flags.
-    pub link_dests: Vec<PathBuf>,
-    /// Enables `--remove-source-files`.
-    pub remove_source_files: bool,
-    /// Optional `--append`/`--no-append` toggle.
-    pub append: Option<bool>,
-    /// Enables `--append-verify`.
-    pub append_verify: bool,
-    /// Optional `--inplace`/`--no-inplace` toggle.
-    pub inplace: Option<bool>,
-    /// Routes daemon messages to standard error via `--msgs2stderr`.
-    pub msgs_to_stderr: bool,
-    /// Optional `--whole-file`/`--no-whole-file` toggle.
-    pub whole_file: Option<bool>,
-    /// Optional bandwidth limit forwarded through `--bwlimit`.
-    ///
-    /// Values are normalised to their rounded byte-per-second representation so
-    /// legacy fallback binaries that predate burst support continue accepting
-    /// the argument. An unlimited transfer is represented using `"0"`.
-    pub bwlimit: Option<OsString>,
-    /// Patterns forwarded via repeated `--exclude` flags.
-    pub excludes: Vec<OsString>,
-    /// Patterns forwarded via repeated `--include` flags.
-    pub includes: Vec<OsString>,
-    /// File paths forwarded via repeated `--exclude-from` flags.
-    pub exclude_from: Vec<OsString>,
-    /// File paths forwarded via repeated `--include-from` flags.
-    pub include_from: Vec<OsString>,
-    /// Raw filter directives forwarded via repeated `--filter` flags.
-    pub filters: Vec<OsString>,
-    /// Number of times the `-F` shortcut was supplied.
-    pub rsync_filter_shortcuts: u8,
-    /// Reference directories forwarded via repeated `--compare-dest` flags.
-    pub compare_destinations: Vec<OsString>,
-    /// Reference directories forwarded via repeated `--copy-dest` flags.
-    pub copy_destinations: Vec<OsString>,
-    /// Reference directories forwarded via repeated `--link-dest` flags.
-    pub link_destinations: Vec<OsString>,
-    /// Enables `--cvs-exclude` on the fallback binary.
-    pub cvs_exclude: bool,
-    /// Values forwarded to the fallback binary via repeated `--info=FLAGS` occurrences.
-    pub info_flags: Vec<OsString>,
-    /// Values forwarded to the fallback binary via repeated `--debug=FLAGS` occurrences.
-    pub debug_flags: Vec<OsString>,
-    /// Whether the original invocation used `--files-from`.
-    pub files_from_used: bool,
-    /// Entries collected from `--files-from` operands.
-    pub file_list_entries: Vec<OsString>,
-    /// Indicates that `--from0` was supplied.
-    pub from0: bool,
-    /// Optional path provided via `--password-file`.
-    pub password_file: Option<PathBuf>,
-    /// Optional daemon password supplied via `--password-file=-`.
-    ///
-    /// When populated the helper writes the password to the fallback
-    /// process' standard input so callers do not need to re-enter
-    /// credentials after the CLI has already consumed them.
-    pub daemon_password: Option<Vec<u8>>,
-    /// Optional protocol override forwarded via `--protocol`.
-    pub protocol: Option<ProtocolVersion>,
-    /// Timeout applied to the spawned process via `--timeout`.
-    pub timeout: TransferTimeout,
-    /// Connection timeout forwarded via `--contimeout`.
-    pub connect_timeout: TransferTimeout,
-    /// Optional `--out-format` template.
-    pub out_format: Option<OsString>,
-    /// Enables `--no-motd`.
-    pub no_motd: bool,
-    /// Preferred address family forwarded via `--ipv4`/`--ipv6`.
-    pub address_mode: AddressMode,
-    /// Optional override for the fallback executable path.
-    ///
-    /// When unspecified the helper consults the
-    /// [`crate::fallback::CLIENT_FALLBACK_ENV`] environment variable
-    /// (`OC_RSYNC_FALLBACK`) and
-    /// defaults to `rsync` if the override is missing or empty.
-    pub fallback_binary: Option<OsString>,
-    /// Optional override for the remote rsync executable.
-    ///
-    /// When populated the helper forwards `--rsync-path` to the fallback command so upstream
-    /// rsync executes the specified program on the remote system. The option is ignored when
-    /// remote operands are absent because local transfers never invoke the fallback binary.
-    pub rsync_path: Option<OsString>,
-    /// Remaining operands to forward to the fallback binary.
-    pub remainder: Vec<OsString>,
-    /// Controls ACL forwarding (`--acls`/`--no-acls`).
-    #[cfg(feature = "acl")]
-    pub acls: Option<bool>,
-    /// Controls xattr forwarding (`--xattrs`/`--no-xattrs`).
-    #[cfg(feature = "xattr")]
-    pub xattrs: Option<bool>,
-}
-
-/// Writer references and arguments required to invoke the fallback binary.
-pub struct RemoteFallbackContext<'a, Out, Err>
-where
-    Out: Write + 'a,
-    Err: Write + 'a,
-{
-    stdout: &'a mut Out,
-    stderr: &'a mut Err,
-    args: RemoteFallbackArgs,
-}
-
-impl<'a, Out, Err> RemoteFallbackContext<'a, Out, Err>
-where
-    Out: Write + 'a,
-    Err: Write + 'a,
-{
-    /// Creates a new context that streams output into the supplied writers.
-    #[must_use]
-    pub fn new(stdout: &'a mut Out, stderr: &'a mut Err, args: RemoteFallbackArgs) -> Self {
-        Self {
-            stdout,
-            stderr,
-            args,
-        }
-    }
-
-    fn split(self) -> (&'a mut Out, &'a mut Err, RemoteFallbackArgs) {
-        let Self {
-            stdout,
-            stderr,
-            args,
-        } = self;
-        (stdout, stderr, args)
-    }
-}
-
-/// Spawns the fallback `rsync` binary with arguments derived from [`RemoteFallbackArgs`].
-///
-/// The helper forwards the subprocess stdout/stderr into the provided writers and returns
-/// the exit status code on success. Errors surface as [`ClientError`] instances with
-/// fully formatted diagnostics.
-pub fn run_remote_transfer_fallback<Out, Err>(
-    stdout: &mut Out,
-    stderr: &mut Err,
-    args: RemoteFallbackArgs,
-) -> Result<i32, ClientError>
-where
-    Out: Write,
-    Err: Write,
-{
-    let RemoteFallbackArgs {
-        dry_run,
-        list_only,
-        remote_shell,
-        remote_options,
-        connect_program,
-        port,
-        bind_address,
-        protect_args,
-        human_readable: human_readable_mode,
-        archive,
-        delete,
-        delete_mode,
-        delete_excluded,
-        max_delete,
-        min_size,
-        max_size,
-        checksum,
-        checksum_choice,
-        checksum_seed,
-        size_only,
-        ignore_existing,
-        ignore_missing_args,
-        update,
-        modify_window,
-        compress,
-        compress_disabled,
-        compress_level,
-        skip_compress,
-        chown,
-        owner,
-        group,
-        chmod,
-        perms,
-        super_mode,
-        times,
-        omit_dir_times,
-        omit_link_times,
-        numeric_ids,
-        hard_links,
-        copy_links,
-        copy_dirlinks,
-        copy_unsafe_links,
-        keep_dirlinks,
-        safe_links,
-        sparse,
-        devices,
-        specials,
-        relative,
-        one_file_system,
-        implied_dirs,
-        mkpath,
-        prune_empty_dirs,
-        verbosity,
-        progress,
-        stats,
-        itemize_changes,
-        partial,
-        preallocate,
-        delay_updates,
-        partial_dir,
-        temp_directory,
-        backup,
-        backup_dir,
-        backup_suffix,
-        link_dests,
-        remove_source_files,
-        append,
-        append_verify,
-        inplace,
-        msgs_to_stderr,
-        whole_file,
-        bwlimit,
-        excludes,
-        includes,
-        exclude_from,
-        include_from,
-        filters,
-        rsync_filter_shortcuts,
-        compare_destinations,
-        copy_destinations,
-        link_destinations,
-        cvs_exclude,
-        info_flags,
-        debug_flags,
-        files_from_used,
-        file_list_entries,
-        from0,
-        password_file,
-        mut daemon_password,
-        protocol,
-        timeout,
-        connect_timeout,
-        out_format,
-        no_motd,
-        address_mode,
-        fallback_binary,
-        rsync_path,
-        mut remainder,
-        #[cfg(feature = "acl")]
-        acls,
-        #[cfg(feature = "xattr")]
-        xattrs,
-    } = args;
-
-    let mut command_args = Vec::new();
-    if archive {
-        command_args.push(OsString::from("-a"));
-    }
-    if dry_run {
-        command_args.push(OsString::from("--dry-run"));
-    }
-    if list_only {
-        command_args.push(OsString::from("--list-only"));
-    }
-    if delete {
-        command_args.push(OsString::from("--delete"));
-        match delete_mode {
-            DeleteMode::Before => command_args.push(OsString::from("--delete-before")),
-            DeleteMode::After => command_args.push(OsString::from("--delete-after")),
-            DeleteMode::Delay => command_args.push(OsString::from("--delete-delay")),
-            DeleteMode::During => command_args.push(OsString::from("--delete-during")),
-            DeleteMode::Disabled => {}
-        }
-    }
-    if delete_excluded {
-        command_args.push(OsString::from("--delete-excluded"));
-    }
-    if backup {
-        command_args.push(OsString::from("--backup"));
-    }
-    if let Some(dir) = backup_dir {
-        command_args.push(OsString::from("--backup-dir"));
-        command_args.push(dir.into_os_string());
-    }
-    if let Some(suffix) = backup_suffix {
-        command_args.push(OsString::from("--suffix"));
-        command_args.push(suffix);
-    }
-    if let Some(limit) = max_delete {
-        let mut arg = OsString::from("--max-delete=");
-        arg.push(limit.to_string());
-        command_args.push(arg);
-    }
-    if let Some(spec) = min_size {
-        let mut arg = OsString::from("--min-size=");
-        arg.push(spec);
-        command_args.push(arg);
-    }
-    if let Some(spec) = max_size {
-        let mut arg = OsString::from("--max-size=");
-        arg.push(spec);
-        command_args.push(arg);
-    }
-    if checksum {
-        command_args.push(OsString::from("--checksum"));
-    }
-    if let Some(choice) = checksum_choice {
-        let mut arg = OsString::from("--checksum-choice=");
-        arg.push(choice);
-        command_args.push(arg);
-    }
-    if let Some(seed) = checksum_seed {
-        let mut arg = OsString::from("--checksum-seed=");
-        arg.push(seed.to_string());
-        command_args.push(arg);
-    }
-    if size_only {
-        command_args.push(OsString::from("--size-only"));
-    }
-    if ignore_existing {
-        command_args.push(OsString::from("--ignore-existing"));
-    }
-    if ignore_missing_args {
-        command_args.push(OsString::from("--ignore-missing-args"));
-    }
-    if update {
-        command_args.push(OsString::from("--update"));
-    }
-    if let Some(window) = modify_window {
-        let mut arg = OsString::from("--modify-window=");
-        arg.push(window.to_string());
-        command_args.push(arg);
-    }
-    if compress {
-        command_args.push(OsString::from("--compress"));
-    } else if compress_disabled {
-        command_args.push(OsString::from("--no-compress"));
-        if whole_file.is_none() {
-            command_args.push(OsString::from("--no-whole-file"));
-        }
-    }
-    if let Some(level) = compress_level {
-        command_args.push(OsString::from("--compress-level"));
-        command_args.push(level);
-    }
-
-    if let Some(spec) = skip_compress {
-        let mut arg = OsString::from("--skip-compress=");
-        arg.push(spec);
-        command_args.push(arg);
-    }
-
-    if let Some(spec) = chown {
-        let mut arg = OsString::from("--chown=");
-        arg.push(spec);
-        command_args.push(arg);
-    }
-
-    push_toggle(&mut command_args, "--owner", "--no-owner", owner);
-    push_toggle(&mut command_args, "--group", "--no-group", group);
-    for spec in chmod {
-        let mut arg = OsString::from("--chmod=");
-        arg.push(&spec);
-        command_args.push(arg);
-    }
-    push_toggle(&mut command_args, "--perms", "--no-perms", perms);
-    push_toggle(&mut command_args, "--super", "--no-super", super_mode);
-    push_toggle(&mut command_args, "--times", "--no-times", times);
-    push_toggle(
-        &mut command_args,
-        "--omit-dir-times",
-        "--no-omit-dir-times",
-        omit_dir_times,
-    );
-    push_toggle(
-        &mut command_args,
-        "--omit-link-times",
-        "--no-omit-link-times",
-        omit_link_times,
-    );
-    push_toggle(
-        &mut command_args,
-        "--numeric-ids",
-        "--no-numeric-ids",
-        numeric_ids,
-    );
-    push_toggle(
-        &mut command_args,
-        "--hard-links",
-        "--no-hard-links",
-        hard_links,
-    );
-    push_toggle(
-        &mut command_args,
-        "--copy-links",
-        "--no-copy-links",
-        copy_links,
-    );
-    if copy_dirlinks {
-        command_args.push(OsString::from("--copy-dirlinks"));
-    }
-    push_toggle(
-        &mut command_args,
-        "--copy-unsafe-links",
-        "--no-copy-unsafe-links",
-        copy_unsafe_links,
-    );
-    push_toggle(
-        &mut command_args,
-        "--keep-dirlinks",
-        "--no-keep-dirlinks",
-        keep_dirlinks,
-    );
-    if safe_links {
-        command_args.push(OsString::from("--safe-links"));
-    }
-    push_toggle(&mut command_args, "--sparse", "--no-sparse", sparse);
-    push_toggle(&mut command_args, "--devices", "--no-devices", devices);
-    push_toggle(&mut command_args, "--specials", "--no-specials", specials);
-    push_toggle(&mut command_args, "--relative", "--no-relative", relative);
-    push_toggle(
-        &mut command_args,
-        "--one-file-system",
-        "--no-one-file-system",
-        one_file_system,
-    );
-    push_toggle(
-        &mut command_args,
-        "--implied-dirs",
-        "--no-implied-dirs",
-        implied_dirs,
-    );
-    if mkpath {
-        command_args.push(OsString::from("--mkpath"));
-    }
-    push_toggle(
-        &mut command_args,
-        "--prune-empty-dirs",
-        "--no-prune-empty-dirs",
-        prune_empty_dirs,
-    );
-    push_toggle(&mut command_args, "--inplace", "--no-inplace", inplace);
-    #[cfg(feature = "acl")]
-    push_toggle(&mut command_args, "--acls", "--no-acls", acls);
-    push_toggle(
-        &mut command_args,
-        "--whole-file",
-        "--no-whole-file",
-        whole_file,
-    );
-    #[cfg(feature = "xattr")]
-    push_toggle(&mut command_args, "--xattrs", "--no-xattrs", xattrs);
-
-    for _ in 0..verbosity {
-        command_args.push(OsString::from("-v"));
-    }
-    if progress {
-        command_args.push(OsString::from("--progress"));
-    }
-    if stats {
-        command_args.push(OsString::from("--stats"));
-    }
-    if itemize_changes {
-        command_args.push(OsString::from("--itemize-changes"));
-    }
-    if partial {
-        command_args.push(OsString::from("--partial"));
-    }
-    if preallocate {
-        command_args.push(OsString::from("--preallocate"));
-    }
-    if delay_updates {
-        command_args.push(OsString::from("--delay-updates"));
-    }
-    if let Some(dir) = partial_dir {
-        command_args.push(OsString::from("--partial-dir"));
-        command_args.push(dir.into_os_string());
-    }
-    if let Some(dir) = temp_directory {
-        command_args.push(OsString::from("--temp-dir"));
-        command_args.push(dir.into_os_string());
-    }
-    for dir in link_dests {
-        let mut arg = OsString::from("--link-dest=");
-        arg.push(dir);
-        command_args.push(arg);
-    }
-    if remove_source_files {
-        command_args.push(OsString::from("--remove-source-files"));
-    }
-    if append_verify {
-        command_args.push(OsString::from("--append-verify"));
-    } else {
-        push_toggle(&mut command_args, "--append", "--no-append", append);
-    }
-    if msgs_to_stderr {
-        command_args.push(OsString::from("--msgs2stderr"));
-    }
-
-    if let Some(enabled) = protect_args {
-        let flag = if enabled {
-            "--protect-args"
-        } else {
-            "--no-protect-args"
-        };
-        command_args.push(OsString::from(flag));
-    }
-
-    push_human_readable(&mut command_args, human_readable_mode);
-
-    if let Some(limit) = bwlimit {
-        command_args.push(OsString::from("--bwlimit"));
-        command_args.push(limit);
-    }
-
-    if let Some(format) = out_format {
-        command_args.push(OsString::from("--out-format"));
-        command_args.push(format);
-    }
-
-    for exclude in excludes {
-        command_args.push(OsString::from("--exclude"));
-        command_args.push(exclude);
-    }
-    for include in includes {
-        command_args.push(OsString::from("--include"));
-        command_args.push(include);
-    }
-    for path in exclude_from {
-        command_args.push(OsString::from("--exclude-from"));
-        command_args.push(path);
-    }
-    for path in include_from {
-        command_args.push(OsString::from("--include-from"));
-        command_args.push(path);
-    }
-    if cvs_exclude {
-        command_args.push(OsString::from("--cvs-exclude"));
-    }
-    for _ in 0..rsync_filter_shortcuts {
-        command_args.push(OsString::from("-F"));
-    }
-    for filter in filters {
-        command_args.push(OsString::from("--filter"));
-        command_args.push(filter);
-    }
-
-    for path in compare_destinations {
-        command_args.push(OsString::from("--compare-dest"));
-        command_args.push(path);
-    }
-
-    for path in copy_destinations {
-        command_args.push(OsString::from("--copy-dest"));
-        command_args.push(path);
-    }
-
-    for path in link_destinations {
-        command_args.push(OsString::from("--link-dest"));
-        command_args.push(path);
-    }
-
-    for flag in info_flags {
-        let mut arg = OsString::from("--info=");
-        arg.push(&flag);
-        command_args.push(arg);
-    }
-
-    for flag in debug_flags {
-        let mut arg = OsString::from("--debug=");
-        arg.push(&flag);
-        command_args.push(arg);
-    }
-
-    let files_from_temp =
-        prepare_file_list(&file_list_entries, files_from_used, from0).map_err(|error| {
-            fallback_error(format!(
-                "failed to prepare file list for fallback rsync invocation: {error}"
-            ))
-        })?;
-
-    if let Some(temp) = files_from_temp.as_ref() {
-        command_args.push(OsString::from("--files-from"));
-        command_args.push(temp.path().as_os_str().to_os_string());
-        if from0 {
-            command_args.push(OsString::from("--from0"));
-        }
-    }
-
-    if let Some(path) = password_file {
-        command_args.push(OsString::from("--password-file"));
-        command_args.push(path.into_os_string());
-    }
-
-    if let Some(protocol) = protocol {
-        command_args.push(OsString::from("--protocol"));
-        command_args.push(OsString::from(protocol.to_string()));
-    }
-
-    match timeout {
-        TransferTimeout::Default => {}
-        TransferTimeout::Disabled => {
-            command_args.push(OsString::from("--timeout"));
-            command_args.push(OsString::from("0"));
-        }
-        TransferTimeout::Seconds(value) => {
-            command_args.push(OsString::from("--timeout"));
-            command_args.push(OsString::from(value.get().to_string()));
-        }
-    }
-
-    match connect_timeout {
-        TransferTimeout::Default => {}
-        TransferTimeout::Disabled => {
-            command_args.push(OsString::from("--contimeout"));
-            command_args.push(OsString::from("0"));
-        }
-        TransferTimeout::Seconds(value) => {
-            command_args.push(OsString::from("--contimeout"));
-            command_args.push(OsString::from(value.get().to_string()));
-        }
-    }
-
-    if no_motd {
-        command_args.push(OsString::from("--no-motd"));
-    }
-
-    for option in remote_options {
-        command_args.push(OsString::from("--remote-option"));
-        command_args.push(option);
-    }
-
-    if let Some(program) = connect_program {
-        command_args.push(OsString::from("--connect-program"));
-        command_args.push(program);
-    }
-
-    if let Some(shell) = remote_shell {
-        command_args.push(OsString::from("-e"));
-        command_args.push(shell);
-    }
-
-    match address_mode {
-        AddressMode::Default => {}
-        AddressMode::Ipv4 => command_args.push(OsString::from("--ipv4")),
-        AddressMode::Ipv6 => command_args.push(OsString::from("--ipv6")),
-    }
-
-    if let Some(port) = port {
-        let mut arg = OsString::from("--port=");
-        arg.push(port.to_string());
-        command_args.push(arg);
-    }
-
-    if let Some(address) = bind_address {
-        let mut arg = OsString::from("--address=");
-        arg.push(address);
-        command_args.push(arg);
-    }
-
-    if let Some(path) = rsync_path {
-        command_args.push(OsString::from("--rsync-path"));
-        command_args.push(path);
-    }
-
-    command_args.append(&mut remainder);
-
-    let binary = if let Some(path) = fallback_binary {
-        path
-    } else {
-        match fallback_override(CLIENT_FALLBACK_ENV) {
-            Some(FallbackOverride::Disabled) => {
-                return Err(fallback_error(format!(
-                    "remote transfers are unavailable because {env} is disabled; set {env} to point to an upstream rsync binary",
-                    env = CLIENT_FALLBACK_ENV
-                )));
-            }
-            Some(other) => other
-                .resolve_or_default(OsStr::new("rsync"))
-                .unwrap_or_else(|| OsString::from("rsync")),
-            None => OsString::from("rsync"),
-        }
-    };
-
-    let mut command = Command::new(&binary);
-    command.args(&command_args);
-    if daemon_password.is_some() {
-        command.stdin(Stdio::piped());
-    } else {
-        command.stdin(Stdio::inherit());
-    }
-    command.stdout(Stdio::piped());
-    command.stderr(Stdio::piped());
-
-    let mut child = command.spawn().map_err(|error| {
-        fallback_error(format!(
-            "failed to launch fallback rsync binary '{}': {error}",
-            Path::new(&binary).display()
-        ))
-    })?;
-
-    if let Some(mut password) = daemon_password.take() {
-        let mut stdin = child
-            .stdin
-            .take()
-            .ok_or_else(|| fallback_error("fallback rsync did not expose a writable stdin"))?;
-
-        write_daemon_password(&mut stdin, &mut password).map_err(|error| {
-            fallback_error(format!(
-                "failed to write password to fallback rsync stdin: {error}"
-            ))
-        })?;
-    }
-
-    let (sender, receiver) = mpsc::channel();
-    let mut stdout_thread = child
-        .stdout
-        .take()
-        .map(|handle| spawn_fallback_reader(handle, FallbackStreamKind::Stdout, sender.clone()));
-    let mut stderr_thread = child
-        .stderr
-        .take()
-        .map(|handle| spawn_fallback_reader(handle, FallbackStreamKind::Stderr, sender.clone()));
-    drop(sender);
-
-    let mut stdout_open = stdout_thread.is_some();
-    let mut stderr_open = stderr_thread.is_some();
-
-    while stdout_open || stderr_open {
-        match receiver.recv() {
-            Ok(FallbackStreamMessage::Data(FallbackStreamKind::Stdout, data)) => {
-                if let Err(error) = stdout.write_all(&data) {
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return Err(fallback_error(format!(
-                        "failed to forward fallback stdout: {error}"
-                    )));
-                }
-            }
-            Ok(FallbackStreamMessage::Data(FallbackStreamKind::Stderr, data)) => {
-                if let Err(error) = stderr.write_all(&data) {
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return Err(fallback_error(format!(
-                        "failed to forward fallback stderr: {error}"
-                    )));
-                }
-            }
-            Ok(FallbackStreamMessage::Error(FallbackStreamKind::Stdout, error)) => {
-                terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                return Err(fallback_error(format!(
-                    "failed to read stdout from fallback rsync: {error}"
-                )));
-            }
-            Ok(FallbackStreamMessage::Error(FallbackStreamKind::Stderr, error)) => {
-                terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                return Err(fallback_error(format!(
-                    "failed to read stderr from fallback rsync: {error}"
-                )));
-            }
-            Ok(FallbackStreamMessage::Finished(kind)) => match kind {
-                FallbackStreamKind::Stdout => stdout_open = false,
-                FallbackStreamKind::Stderr => stderr_open = false,
-            },
-            Err(_) => {
-                if stdout_open {
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return Err(fallback_error(
-                        "failed to capture stdout from fallback rsync binary",
-                    ));
-                }
-                if stderr_open {
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return Err(fallback_error(
-                        "failed to capture stderr from fallback rsync binary",
-                    ));
-                }
-                break;
-            }
-        }
-    }
-
-    join_fallback_thread(&mut stdout_thread);
-    join_fallback_thread(&mut stderr_thread);
-
-    let status = child.wait().map_err(|error| {
-        fallback_error(format!(
-            "failed to wait for fallback rsync process: {error}"
-        ))
-    })?;
-
-    drop(files_from_temp);
-
-    Ok(status.code().unwrap_or(MAX_EXIT_CODE))
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum FallbackStreamKind {
-    Stdout,
-    Stderr,
-}
-
-enum FallbackStreamMessage {
-    Data(FallbackStreamKind, Vec<u8>),
-    Error(FallbackStreamKind, io::Error),
-    Finished(FallbackStreamKind),
-}
-
-fn fallback_error(text: impl Into<String>) -> ClientError {
-    let message = rsync_error!(1, "{}", text.into()).with_role(Role::Client);
-    ClientError::new(1, message)
-}
-
-fn push_toggle(args: &mut Vec<OsString>, enable: &str, disable: &str, setting: Option<bool>) {
-    match setting {
-        Some(true) => args.push(OsString::from(enable)),
-        Some(false) => args.push(OsString::from(disable)),
-        None => {}
-    }
-}
-
-fn push_human_readable(args: &mut Vec<OsString>, mode: Option<HumanReadableMode>) {
-    match mode {
-        Some(HumanReadableMode::Disabled) => args.push(OsString::from("--no-human-readable")),
-        Some(HumanReadableMode::Enabled) => args.push(OsString::from("--human-readable")),
-        Some(HumanReadableMode::Combined) => args.push(OsString::from("--human-readable=2")),
-        None => {}
-    }
-}
-
-fn prepare_file_list(
-    entries: &[OsString],
-    files_from_used: bool,
-    zero_terminated: bool,
-) -> io::Result<Option<NamedTempFile>> {
-    if !files_from_used {
-        return Ok(None);
-    }
-
-    let mut file = NamedTempFile::new()?;
-    {
-        let writer = file.as_file_mut();
-        for entry in entries {
-            write_file_list_entry(writer, entry.as_os_str())?;
-            if zero_terminated {
-                writer.write_all(&[0])?;
-            } else {
-                writer.write_all(b"\n")?;
-            }
-        }
-        writer.flush()?;
-    }
-
-    Ok(Some(file))
-}
-
-fn write_file_list_entry<W: Write>(writer: &mut W, value: &OsStr) -> io::Result<()> {
-    #[cfg(unix)]
-    {
-        writer.write_all(value.as_bytes())
-    }
-
-    #[cfg(not(unix))]
-    {
-        writer.write_all(value.to_string_lossy().as_bytes())
-    }
-}
-
-fn spawn_fallback_reader<R>(
-    mut reader: R,
-    kind: FallbackStreamKind,
-    sender: Sender<FallbackStreamMessage>,
-) -> thread::JoinHandle<()>
-where
-    R: Read + Send + 'static,
-{
-    thread::spawn(move || {
-        let mut buffer = [0u8; 8192];
-        loop {
-            match reader.read(&mut buffer) {
-                Ok(0) => {
-                    let _ = sender.send(FallbackStreamMessage::Finished(kind));
-                    break;
-                }
-                Ok(n) => {
-                    if sender
-                        .send(FallbackStreamMessage::Data(kind, Vec::from(&buffer[..n])))
-                        .is_err()
-                    {
-                        break;
-                    }
-                }
-                Err(error) => {
-                    let _ = sender.send(FallbackStreamMessage::Error(kind, error));
-                    break;
-                }
-            }
-        }
-    })
-}
-
-/// Writes the daemon password into `writer`, appending a newline when required and
-/// scrubbing the buffer afterwards.
-fn write_daemon_password<W: Write>(writer: &mut W, password: &mut Vec<u8>) -> io::Result<()> {
-    if !password.ends_with(b"\n") {
-        password.push(b'\n');
-    }
-
-    writer.write_all(password)?;
-    writer.flush()?;
-
-    for byte in password.iter_mut() {
-        *byte = 0;
-    }
-
-    Ok(())
-}
-
-fn join_fallback_thread(handle: &mut Option<thread::JoinHandle<()>>) {
-    if let Some(join_handle) = handle.take() {
-        let _ = join_handle.join();
-    }
-}
-
-fn terminate_fallback_process(
-    child: &mut Child,
-    stdout_thread: &mut Option<thread::JoinHandle<()>>,
-    stderr_thread: &mut Option<thread::JoinHandle<()>>,
-) {
-    let _ = child.kill();
-    let _ = child.wait();
-    join_fallback_thread(stdout_thread);
-    join_fallback_thread(stderr_thread);
-}
-
 /// Describes a transfer action performed by the local copy engine.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub enum ClientEventKind {
@@ -4799,6 +3728,8 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::client::fallback::write_daemon_password;
+    use crate::fallback::CLIENT_FALLBACK_ENV;
     use crate::version::RUST_VERSION;
     use rsync_compress::zlib::CompressionLevel;
     use std::ffi::{OsStr, OsString};
@@ -4807,7 +3738,7 @@ mod tests {
     use std::net::{Shutdown, TcpListener, TcpStream};
     use std::num::{NonZeroU8, NonZeroU64};
     use std::path::{Path, PathBuf};
-    use std::sync::{Mutex, OnceLock};
+    use std::sync::{Mutex, OnceLock, mpsc};
     use std::thread;
     use std::time::Duration;
     use tempfile::tempdir;
